+++ src/test/java/kitchenpos/menu/acceptance/MenuGroupAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance;
+package kitchenpos.menu.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,7 @@ import io.restassured.response.Response;
 import java.util.List;
 import kitchenpos.AcceptanceTest;
 import kitchenpos.acceptance.util.KitchenPosBehaviors;
+import kitchenpos.menu.acceptance.behavior.MenuContextBehavior;
 import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,10 +20,10 @@ class MenuGroupAcceptanceTest extends AcceptanceTest {
     @Test
     @DisplayName("메뉴그룹 생성 및 조회 기능 인수테스트")
     void menuGroupAcceptanceTest() {
-        ExtractableResponse<Response> createResponse = KitchenPosBehaviors.메뉴그룹_생성_요청("치킨");
+        ExtractableResponse<Response> createResponse = MenuContextBehavior.메뉴그룹_생성_요청("치킨");
         assertThat(createResponse.statusCode()).isEqualTo(HttpStatus.CREATED.value());
 
-        List<MenuGroup> menuGroups = KitchenPosBehaviors.메뉴그룹_목록조회();
+        List<MenuGroup> menuGroups = MenuContextBehavior.메뉴그룹_목록조회();
         assertThat(menuGroups).hasSize(1);
     }
 }
